fix: Organize codes of pages clearly
