+++ frontend/app/(auth)/input-otp/from-login/page.tsx
@@ -1,39 +1,30 @@
 "use client";
-//import from react/next
-import { useState, useEffect } from "react";
-import Image from "next/image";
+
 import Link from "next/link";
+import Image from "next/image";
+import { useState, useEffect } from "react";
 import { useRouter } from "next/navigation";
 
-// import from antd
 import { Flex, Typography, Button } from "antd";
 
-// import from components
+import WETPOTPInput from "@/components/auth/input/WETPOTPInput";
 import SpinLoading from "@/components/general/SpinLoading";
-import OTPInput from "@/components/auth/input/OTPInput";
 import AuthHeader from "@/components/auth/header/Header-auth";
 
-// import images
 import SendRequestIcon from "@/public/images/auth/SendRequest.svg";
 import BackArrowIcon from "@/public/images/auth/arrow-ios-back-fill.svg";
-
-// import from other modules
 
 export default function Page() {
   const router = useRouter();
 
-  // Judge whether page loaded
   const [loaded, setLoaded] = useState<boolean>(false);
+  const [isVerified, setIsVerified] = useState<boolean>(false);
+  const [isMobile, setIsMobile] = useState<boolean | null>(null);
 
   useEffect(() => {
     setTimeout(() => setLoaded(true), 1);
   }, []);
 
-  // Judge whether mobile view
-  const [isMobile, setIsMobile] = useState<boolean | null>(null);
-  const handleResize = () => {
-    setIsMobile(window.innerWidth < 1024);
-  };
   useEffect(() => {
     handleResize();
     window.addEventListener("resize", handleResize);
@@ -42,13 +33,14 @@ export default function Page() {
     };
   }, []);
 
-  // Judge whether success or failed
-  const [isVerified, setIsVerified] = useState<boolean>(false);
+  const handleResize = () => {
+    setIsMobile(window.innerWidth < 1024);
+  };
 
-  // event handlers
   const onClickVerifyOTPHandler = () => {
     router.push("/reset-password");
   };
+
   const onClickCancelHandler = () => {
     router.push("/forgot-password");
   };
@@ -80,7 +72,7 @@ export default function Page() {
               </Typography>
             </Flex>
             <Flex justify="center" className={`w-full px-4`}>
-              <OTPInput />
+              <WETPOTPInput />
             </Flex>
             <Flex vertical className="w-full px-4 gap-y-6">
               <Flex vertical className="w-full gap-y-3">
