+++ frontend/components/auth/header/Header-sign.tsx
@@ -7,7 +7,7 @@ import { Flex, Typography } from "antd";
 // import images
 import Logomark from "@/public/images/auth/Logomark.svg";
 
-export default function Header() {
+export default function SignHeader() {
   return (
     <Flex // Header
       className="w-full h-11 lg:h-[72px] bg-main/[12%] flex justify-center lg:justify-start items-center"
